created a JokeController to control the flow of the datamodel.
